Refatorando o código e criando comandos para consulta
