+++ cypress/e2e/consulta.cy.js
@@ -1,5 +1,3 @@
-require('cypress-xpath')
-
 describe('Testes de Agendar Consulta', () => {
 
     beforeEach(() => {
@@ -7,79 +5,28 @@ describe('Testes de Agendar Consulta', () => {
     })
 
     it('Validar agendamento com horário disponível', () => {
-        const tomorrow = new Date();
-        tomorrow.setDate(tomorrow.getDate() + 1)
-        const yyyy = tomorrow.getFullYear()
-        const mm = String(tomorrow.getMonth() + 1).padStart(2, '0')
-        const dd = String(tomorrow.getDate()).padStart(2, '0')
-        const dateStr = `${yyyy}-${mm}-${dd}`
-
-        cy.get('#pacienteSelect').parent().find('input.select-dropdown').click({ force: true })
-        cy.xpath("//span[text()='João Pedro Almeida (123.456.789-00)']").click({ force: true })
-        cy.get('#profissionalSelect').parent().find('input.select-dropdown').click({ force: true })
-        cy.xpath("//span[text()='Dra. Ana Paula Santos — Endodontia']").click({ force: true })
-        cy.get('#dataInput').type(dateStr)
-        cy.get(':nth-child(4) > :nth-child(2) > .select-wrapper > .dropdown-trigger').click()
-        cy.get(':nth-child(4) > :nth-child(2) > .select-wrapper > .dropdown-trigger').click()
-        cy.xpath(`//span[text()='10:00']`).click()
-        cy.get('#tipoConsultaInput').type('Limpeza')
-        cy.get('#obsInput').type('Teste automatizado')
-        cy.get('#btnAgendar').click()
-        cy.get('.toast').should('have.text', 'Consulta agendada!')
+        cy.fazerAAgendamentoDeConsultaComDadosValidos()
+        cy.validarMensagem('Consulta agendada!')
     })
-    
+
     it('Validar agendamento sem informar o profissional', () => {
-        const tomorrow = new Date();
-        tomorrow.setDate(tomorrow.getDate() + 1)
-        const yyyy = tomorrow.getFullYear()
-        const mm = String(tomorrow.getMonth() + 1).padStart(2, '0')
-        const dd = String(tomorrow.getDate()).padStart(2, '0')
-        const dateStr = `${yyyy}-${mm}-${dd}`
+        cy.fazerAAgendamentoDeConsultaSemPreencherOProfissional()
+        cy.validarMensagem('Preencha todos os campos obrigatórios')
+    })
 
-        cy.get('#pacienteSelect').parent().find('input.select-dropdown').click({ force: true })
-        cy.xpath("//span[text()='João Pedro Almeida (123.456.789-00)']").click({ force: true })
-        cy.get('#dataInput').type(dateStr)
-        cy.get('#tipoConsultaInput').type('Limpeza')
-        cy.get('#obsInput').type('Teste automatizado')
-        cy.get('#btnAgendar').click()
-        cy.get('.toast').should('have.text', 'Preencha todos os campos obrigatórios')
-    })
-    
     it('Validar agendamento sem informar o paciente', () => {
-        const tomorrow = new Date();
-        tomorrow.setDate(tomorrow.getDate() + 1)
-        const yyyy = tomorrow.getFullYear()
-        const mm = String(tomorrow.getMonth() + 1).padStart(2, '0')
-        const dd = String(tomorrow.getDate()).padStart(2, '0')
-        const dateStr = `${yyyy}-${mm}-${dd}`
-        cy.get('#profissionalSelect').parent().find('input.select-dropdown').click({ force: true })
-        cy.xpath("//span[text()='Dra. Ana Paula Santos — Endodontia']").click({ force: true })
-        cy.get('#dataInput').type(dateStr)
-        cy.get(':nth-child(4) > :nth-child(2) > .select-wrapper > .dropdown-trigger').click()
-        cy.get(':nth-child(4) > :nth-child(2) > .select-wrapper > .dropdown-trigger').click()
-        cy.xpath(`//span[text()='09:00']`).click()
-        cy.get('#tipoConsultaInput').type('Limpeza')
-        cy.get('#obsInput').type('Teste automatizado')
-        cy.get('#btnAgendar').click()
-        cy.get('.toast').should('have.text', 'Preencha todos os campos obrigatórios')
+        cy.fazerAAgendamentoDeConsultaSemPreencherOPaciente()
+        cy.validarMensagem('Preencha todos os campos obrigatórios')
     })
-    
+
     it('Validar agendamento sem informar o horário', () => {
-        const tomorrow = new Date();
-        tomorrow.setDate(tomorrow.getDate() + 1)
-        const yyyy = tomorrow.getFullYear()
-        const mm = String(tomorrow.getMonth() + 1).padStart(2, '0')
-        const dd = String(tomorrow.getDate()).padStart(2, '0')
-        const dateStr = `${yyyy}-${mm}-${dd}`
-        cy.get('#profissionalSelect').parent().find('input.select-dropdown').click({ force: true })
-        cy.xpath("//span[text()='Dra. Ana Paula Santos — Endodontia']").click({ force: true })
-        cy.get('#dataInput').type(dateStr)
-        cy.get(':nth-child(4) > :nth-child(2) > .select-wrapper > .dropdown-trigger').click()
-        cy.get(':nth-child(4) > :nth-child(2) > .select-wrapper > .dropdown-trigger').click()
-        cy.get('#tipoConsultaInput').type('Limpeza')
-        cy.get('#obsInput').type('Teste automatizado')
-        cy.get('#btnAgendar').click()
-        cy.get('.toast').should('have.text', 'Preencha todos os campos obrigatórios')
+        cy.fazerAAgendamentoDeConsultaSemInformarOHorario()
+        cy.validarMensagem('Preencha todos os campos obrigatórios')
+    })
+
+    it('Validar agendamento sem informar a data', () => {
+        cy.fazerAAgendamentoDeConsultaSemInformarAData()
+        cy.validarMensagem('Preencha todos os campos obrigatórios')
     })
 
 })
